Use minimal __main__.py without the __name__ check according to Python doc
