+++ python/grass/app/Makefile
@@ -7,6 +7,7 @@ DSTDIR = $(ETC)/python/grass/app
 
 MODULES = \
 	__main__ \
+	cli \
 	data \
 	runtime
